fix: threshold was on exact max & min of value, overflow on lower one, add +/-1 to prevent that on threashold removing

+++ include/sos/sosType.hpp
@@ -33,7 +33,7 @@ namespace sos { // namespace start
 /* TYPE */
 
 /* type simplification */
-using Byte = std::uint64_t;
+using Byte = std::uint16_t;
 using Bytes = std::vector<sos::Byte>;
 using Key = Bytes;
